Refactor: MongoDB read로 조회수 이동
