+++ src/main/java/com/playus/communityservice/domain/comment/repository/write/CommentRepository.java
@@ -2,12 +2,10 @@ package com.playus.communityservice.domain.comment.repository.write;
 
 import com.playus.communityservice.domain.comment.entity.Comment;
 import com.playus.communityservice.domain.comment.entity.CommentGroup;
-import com.playus.communityservice.domain.post.entity.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
     List<Comment> findAllByCommentGroup(CommentGroup commentGroup);
-    List<Comment> findAllByCommentGroup_Post(Post post);
 }
