+++ src/main/java/com/playus/communityservice/domain/post/entity/Post.java
@@ -84,7 +84,5 @@ public class Post extends BaseTimeEntity {
     public void delete() {
         this.activated = false;
     }
-
-    public void increaseView() { this.view++; }
 }
