+++ src/main/java/com/playus/communityservice/global/config/data/mongodb/ReadMongoConfig.java
@@ -10,7 +10,6 @@ import org.springframework.data.mongodb.repository.config.EnableMongoRepositorie
 @EnableMongoRepositories(
         basePackages = {
                 "com.playus.communityservice.domain.comment.repository.read",
-                "com.playus.communityservice.domain.file.repository.read",
                 "com.playus.communityservice.domain.post.repository.read"
         },
         mongoTemplateRef = "readMongoTemplate"
